NUnitRetries tests, and refactor.
