+++ Datadog.Trace.OSX.slnf
@@ -112,6 +112,7 @@
       "tracer\\test\\test-applications\\integrations\\Samples.NetActivitySdk\\Samples.NetActivitySdk.csproj",
       "tracer\\test\\test-applications\\integrations\\Samples.NoMultiLoader\\Samples.NoMultiLoader.csproj",
       "tracer\\test\\test-applications\\integrations\\Samples.Npgsql\\Samples.Npgsql.csproj",
+      "tracer\\test\\test-applications\\integrations\\Samples.NUnitTestsRetries\\Samples.NUnitTestsRetries.csproj",
       "tracer\\test\\test-applications\\integrations\\Samples.NUnitTests\\Samples.NUnitTests.csproj",
       "tracer\\test\\test-applications\\integrations\\Samples.OpenTelemetrySdk\\Samples.OpenTelemetrySdk.csproj",
       "tracer\\test\\test-applications\\integrations\\Samples.OracleMDA\\Samples.OracleMDA.csproj",
