+++ Datadog.Trace.sln
@@ -578,6 +578,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Samples.Debugger.AspNetCore
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples.XUnitTestsRetries", "tracer\test\test-applications\integrations\Samples.XUnitTestsRetries\Samples.XUnitTestsRetries.csproj", "{B8DAF87D-A30D-48AE-B630-C65A64D0C3AF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples.NUnitTestsRetries", "tracer\test\test-applications\integrations\Samples.NUnitTestsRetries\Samples.NUnitTestsRetries.csproj", "{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1390,6 +1392,10 @@ Global
 		{B8DAF87D-A30D-48AE-B630-C65A64D0C3AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B8DAF87D-A30D-48AE-B630-C65A64D0C3AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B8DAF87D-A30D-48AE-B630-C65A64D0C3AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1616,6 +1622,7 @@ Global
 		{C4ABF344-3263-45D5-A074-03FB206FF309} = {498A300E-D036-49B7-A43D-821D1CAF11A5}
 		{3978A7D5-7B6E-4152-9C3A-5852F1F6E223} = {16427BFB-B4C6-46A9-A290-8EA51FF73FEA}
 		{B8DAF87D-A30D-48AE-B630-C65A64D0C3AF} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
+		{7131FE5A-6B27-4BBC-B0CF-09780F6D2DFE} = {BAF8F246-3645-42AD-B1D0-0F7EAFBAB34A}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {160A1D00-1F5B-40F8-A155-621B4459D78F}
